requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.3.0
bcrypt==5.0.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.1.0
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.12.0
fastapi==0.115.8
filelock==3.17.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.60.1
fpdf2==2.8.5
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.14
MarkupSafe==3.0.3
narwhals==2.1.2
numpy==2.2.2
openai==1.100.2
openpyxl==3.1.5
orjson==3.11.2
packaging==25.0
pandas==2.2.3
pillow==11.3.0
platformdirs==4.3.6
protobuf==6.32.0
pyarrow==21.0.0
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydyf==0.11.0
PyMySQL==1.1.2
pypdf==6.0.0
pyphen==0.17.2
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.27.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.45.3
streamlit==1.48.1
tenacity==9.1.2
tiktoken==0.11.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
virtualenv==20.29.1
watchdog==6.0.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zopfli==0.2.3.post1
zstandard==0.24.0
